+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@ Changelog
 - Enh #270: Rename soft deleted wiki page automatically on creating new page with same name
 - Enh #277: Fix visibility of "Delete" link on wall entry
 - Fix #274: Fix left indent of sub-pages on default theme
+- Fix #281: Disable "Parent Page" when user has no permission to edit it
 
 2.0.0-beta.1 (July 28, 2023)
 ----------------------------